Add: 기본 이미지 목록 받아오는 엔드포인트 추가 & 컨트롤러 분리 현재 에셋을 못받은 관계로 임시 더미 리턴
