+++ src/main/java/net/mureng/mureng/reply/web/ReplyController.java
@@ -65,17 +65,4 @@ public class ReplyController {
         replyService.delete(member, replyId);
         return ResponseEntity.ok(ApiResult.ok(true));
     }
-
-    @ApiOperation(value = "답변 이미지 업로드", notes = "답변에 적용될 이미지를 업로드합니다.")
-    @PostMapping(value = "/image", consumes = {MediaType.MULTIPART_FORM_DATA_VALUE})
-    public ResponseEntity<ApiResult<ImagePathDto>> postImage(
-            @RequestParam("image") MultipartFile image) {
-        return ResponseEntity.ok(ApiResult.ok(new ImagePathDto(replyService.uploadReplyImageFile(image))));
-    }
-
-    @Getter
-    @AllArgsConstructor
-    public static class ImagePathDto {
-        private final String imagePath;
-    }
 }
